requirements.txt
flask
flask_cors
Pillow
numpy
opencv-python-headless==4.2.0.32
cvzone
mediapipe
gunicorn